fix: use canvas package in combination with jsdom
The `jsdom` package states it comes with support for canvas
https://github.com/jsdom/jsdom?tab=readme-ov-file#canvas-support, and
also has the `canvas` npm package listed as a potential peer dependency.

I removed the previous polyfill and added the node `canvas` package to
the dependencies. The one downside obviously is now `excalidraw_export`
expects several binaries to be installed on the system (due to the
`canvas` library needing them).

This fixes https://github.com/Timmmm/excalidraw_export/issues/11.